backend/agents/company_matcher/src/company_matcher/state.py
"""State definitions for Company Matcher."""

from typing import Optional
from langgraph.graph import MessagesState


class CompanyMatcherInputState(MessagesState):
    """Input state - messages containing the company name and country."""
    country_of_establishment: Optional[str] = None


class CompanyMatcherState(MessagesState):
    """State for the company matcher workflow."""
    
    company_name: str = ""
    country_of_establishment: str = ""
    match_result: str = ""

